move ScrollableLabel (#13)
